+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Changed
+
+- Tally and Key Ceremony components API have been changed a bit. The `setupElection` method should be called before binding the UI elements to the component.
+
 ## [0.19.0] - 2021-04-16
 
 ## Added